+++ _src/interfaces/executionContext.ts
@@ -2,4 +2,5 @@ import { ExecutionScope } from './executionScope';
 
 export interface ExecutionContext {
 	scope?: ExecutionScope;
+	source?: string;
 }